pyproject.toml
[tool.poetry]
name = "cxr-generation"
version = "0.1.0"
description = ""
authors = ["Rajat Rasal"]
readme = "README.md"
packages = [
    { include = "dataset", from = "src" },
    { include = "semantic_editing", from = "src" }
]

[tool.poetry.scripts]
caption_imagenet = "dataset.imagenet_captioning:main"
laion_400m_download = "dataset.laion_400m:main"
laion_400m_embed = "dataset.clip:main"
format_reports = "dataset.create_section_files:cli"
experiments = "semantic_editing.experiments:main"
nti = "semantic_editing.null_text_inversion:main"

[tool.poetry.dependencies]
python = "3.9.18"
torch = "2.1.2"
diffusers = "0.26.2"
accelerate = "0.26.1"
pillow = "9.3.0"
numpy = "1.26.3"
blobfile = "1.0.5"
mpi4py = "3.1.5"
setuptools = "68.2.2"
matplotlib = "3.8.2"
pandas = "2.1.4"
tqdm = "4.66.1"
transformers = "4.36.2"
torchvision = "0.16.2"
torchmetrics = "1.3.1"
scikit-learn = "1.4.1.post1"
pytest = "8.0.1"
pytest-dependency = "0.6.0"
litdata = "0.2.3"
autofaiss = "^2.9.6"
clip-retrieval = "2.44.0"
webdataset = "0.2.86"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
bitsandbytes = "0.43.0"
triton = "2.1.0"
nvidia-cuda-nvrtc-cu12 = "12.1.105"
nvidia-cuda-runtime-cu12 = "12.1.105"
nvidia-cuda-cupti-cu12 = "12.1.105"
nvidia-cudnn-cu12 = "8.9.2.26"
nvidia-cublas-cu12 = "12.1.3.1"
nvidia-cufft-cu12 = "11.0.2.54"
nvidia-curand-cu12 = "10.3.2.106"
nvidia-cusolver-cu12 = "11.4.5.107"
nvidia-cusparse-cu12 = "12.1.0.106"
nvidia-nccl-cu12 = "2.18.1"
nvidia-nvtx-cu12 = "12.1.105"
nvidia-nvjitlink-cu12 = "12.3.52"

[build-system]
requires = ["poetry-core==1.7.*"]
build-backend = "poetry.core.masonry.api"
